@@ -1,37 +1,49 @@
 {
-    "name": "zumba/json-serializer",
-    "type": "library",
-    "description": "Serialize PHP variables, including objects, in JSON format. Support to unserialize it too.",
-    "keywords": ["json", "serialize", "serializer"],
-    "homepage": "https://tech.zumba.com",
-    "license": "MIT",
-    "scripts": {
-        "test": "phpunit --colors=always"
-    },
-    "authors": [
-        {
-            "name": "Zumba Fitness, LLC",
-            "email": "engineering@zumba.com"
-        },
-        {
-            "name": "Juan Basso",
-            "email": "juan.basso@zumba.com"
-        }
-    ],
-    "require": {
-        "php": "^7.2 || ^8.0",
-        "ext-mbstring": "*"
-    },
-    "suggest": {
-        "opis/closure": "Allow to serialize PHP closures"
-    },
-    "require-dev": {
-        "phpunit/phpunit": ">=8 <12.0"
-    },
-    "autoload": {
-        "psr-4": {
-            "Zumba\\": "src/",
-            "Zumba\\JsonSerializer\\Test\\": "tests/"
-        }
+  "name": "zumba/json-serializer",
+  "type": "library",
+  "description": "Serialize PHP variables, including objects, in JSON format. Support to unserialize it too.",
+  "keywords": [
+    "json",
+    "serialize",
+    "serializer"
+  ],
+  "homepage": "https://tech.zumba.com",
+  "license": "MIT",
+  "scripts": {
+    "test": "phpunit --colors=always"
+  },
+  "repositories": {
+    "jeremeamia-superclosure-php84": {
+      "type": "github",
+      "url": "git@github.com:Bitcaster/super_closure.git"
     }
+  },
+  "authors": [
+    {
+      "name": "Zumba Fitness, LLC",
+      "email": "engineering@zumba.com"
+    },
+    {
+      "name": "Juan Basso",
+      "email": "juan.basso@zumba.com"
+    }
+  ],
+  "require": {
+    "php": "^8.4",
+    "ext-mbstring": "*",
+    "ext-mongodb": "*",
+    "jeremeamia/superclosure": "dev-php8.4"
+  },
+  "suggest": {
+    "opis/closure": "Allow to serialize PHP closures"
+  },
+  "require-dev": {
+    "phpunit/phpunit": ">=8 <12.0"
+  },
+  "autoload": {
+    "psr-4": {
+      "Zumba\\": "src/",
+      "Zumba\\JsonSerializer\\Test\\": "tests/"
+    }
+  }
 }
