@@ -12,14 +12,15 @@ interface ClosureSerializer {
      * @param Closure $closure
      * @return string
      */
-    public function serialize(Closure $closure);
+    public function serialize(Closure $closure): string;
 
     /**
      * Unserialize a closure
      *
      * @param string $serialized
+     *
      * @return Closure
      */
-    public function unserialize($serialized);
+    public function unserialize(string $serialized): Closure;
 
 }
