@@ -13,7 +13,7 @@ class OpisClosureSerializer implements ClosureSerializer {
      * @param Closure $closure
      * @return string
      */
-    public function serialize(Closure $closure)
+    public function serialize(Closure $closure): string
     {
         return serialize(new OpisSerializableClosure($closure));
     }
@@ -22,9 +22,10 @@ class OpisClosureSerializer implements ClosureSerializer {
      * Unserialize a closure
      *
      * @param string $serialized
+     *
      * @return Closure
      */
-    public function unserialize($serialized)
+    public function unserialize(string $serialized): Closure
     {
         return unserialize($serialized)->getClosure();
     }
